requirements.txt
mysql-connector-python
pymongo
python -m venv venv
venv\Scripts\activate
pip install -r requirements.txt